requirement.txt

### `requirements.txt`

```text
google-adk[a2a]
mcp[cli]
litellm
uvicorn
nest_asyncio
google-genai
python-dotenv
requests